[FPM]: Fix dplane_fpm_sonic for SRv6 SID list with multiple SIDs
Currently, there is a bug in the dplane_fpm_sonic to handle an SRv6 SID
List containing multiple SIDs.
dplane_fpm_sonic is responsible for encoding the route in a Netlink
message and sending it to SONiC fpmsyncd for installation in the ApplDB.

When FRR has a route with an SRv6 nexthop containing a SID List composed
of multiple SIDs, dplane_fpm_sonic only encodes the first SID and omits
the subsequent SIDs.

As a result, SONiC fpmsyncd receives a route with a wrong SID List,
containing only the first SID.

This pull request fixes the issue in dplane_fpm_sonic to handle the
entire SID list in the Netlink message, ensuring the correct SID list
is installed in the ApplDB.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>